#18 🧪 Add Unit tests for cpu module

+++ kamper/modules/cpu/src/commonMain/kotlin/com/smellouk/kamper/cpu/CpuInfo.kt
@@ -3,13 +3,13 @@ package com.smellouk.kamper.cpu
 import com.smellouk.kamper.api.Info
 
 data class CpuInfo(
-    val totalUseRatio: Float,
-    val appRatio: Float,
-    val userRatio: Float,
-    val systemRatio: Float,
-    val ioWaitRatio: Float
+    val totalUseRatio: Double,
+    val appRatio: Double,
+    val userRatio: Double,
+    val systemRatio: Double,
+    val ioWaitRatio: Double
 ) : Info {
     companion object {
-        val INVALID = CpuInfo(-1F, -1F, -1F, -1F, -1F)
+        val INVALID = CpuInfo(-1.0, -1.0, -1.0, -1.0, -1.0)
     }
 }
